Viewmodel reorganisation (#158)
* ViewModel subfolder with separate files for each ViewModelItem
* Use a callable to decouple viewmodel and viewmodel update
* Granular update queue lock
* Use bulk update and removal of lines from batches
* Separate file for each command
* Be less aggressive about saving the project file
* Comments and cleanup
