+++ GUI/Commands/LoadSubtitleFile.py
@@ -1,10 +1,9 @@
-import logging
 from GUI.Command import Command
 from GUI.ProjectDataModel import ProjectDataModel
-from PySubtitle.Helpers import GetOutputPath
 from PySubtitle.Options import Options
-
 from PySubtitle.SubtitleProject import SubtitleProject
+
+import logging
 
 class LoadSubtitleFile(Command):
     def __init__(self, filepath, options : Options):
@@ -37,7 +36,7 @@ class LoadSubtitleFile(Command):
                 self.datamodel.CreateViewModel()
 
             return True
-        
+
         except Exception as e:
             logging.error(f"Unable to load {self.filepath} ({str(e)})")
             return False
@@ -45,38 +44,3 @@ class LoadSubtitleFile(Command):
     def undo(self):
         # I suppose we _could_ store a reference to the previous project...
         pass
-
-class SaveProjectFile(Command):
-    def __init__(self, project : SubtitleProject, filepath = None):
-        super().__init__()
-        self.project = project
-        self.filepath = filepath or project.subtitles.outputpath
-
-    def execute(self):
-        self.project.projectfile = self.project.GetProjectFilepath(self.filepath)
-        self.project.subtitles.outputpath = GetOutputPath(self.project.projectfile, self.project.target_language)
-        self.project.WriteProjectFile()
-
-        if self.project.subtitles.translated:
-            self.project.SaveTranslation()
-        return True
-
-class SaveSubtitleFile(Command):
-    def __init__(self, filepath, project : SubtitleProject):
-        super().__init__()
-        self.filepath = filepath
-        self.project = project
-
-    def execute(self):
-        self.project.subtitles.SaveOriginals(self.filepath)
-        return True
-
-class SaveTranslationFile(Command):
-    def __init__(self, filepath, project : SubtitleProject):
-        super().__init__()
-        self.filepath = filepath
-        self.project = project
-
-    def execute(self):
-        self.project.SaveTranslation(self.filepath)
-        return True
