+++ GUI/SubtitleListModel.py
@@ -1,7 +1,11 @@
 import logging
 from PySide6.QtCore import QAbstractProxyModel, QModelIndex, Qt
-from GUI.ProjectViewModel import BatchItem, ProjectViewModel, SceneItem, LineItem, ViewModelItem
+from GUI.ViewModel.SceneItem import SceneItem
+from GUI.ViewModel.BatchItem import BatchItem
+from GUI.ViewModel.LineItem import LineItem
+from GUI.ViewModel.ViewModel import ProjectViewModel
 from GUI.ProjectSelection import ProjectSelection
+from GUI.ViewModel.ViewModelItem import ViewModelItem
 from GUI.Widgets.Widgets import LineItemView
 
 class SubtitleListModel(QAbstractProxyModel):
