+++ GUI/ViewModel/ViewModelUpdateSection.py
@@ -2,8 +2,8 @@ class ModelUpdateSection:
     def __init__(self):
         self.updates = {}
         self.replacements = {}
-        self.removals = []
         self.additions = {}
+        self.removals = []
 
     def update(self, key, item_update):
         self.updates[key] = item_update
@@ -18,19 +18,8 @@ class ModelUpdateSection:
         self.removals.append(key)
 
     def HasUpdate(self) -> bool:
-        return self.updates or self.replacements or self.additions or self.removals
-    
+        return True if (self.updates or self.replacements or self.additions or self.removals) else False
+
     @property
     def size_changed(self) -> bool:
-        return self.removals or self.additions
-
-class ModelUpdate:
-    def __init__(self):
-        self.rebuild = False
-        self.scenes = ModelUpdateSection()
-        self.batches = ModelUpdateSection()
-        self.lines = ModelUpdateSection()
-
-    def HasUpdate(self) -> bool:
-        return self.scenes.HasUpdate() or self.batches.HasUpdate() or self.lines.HasUpdate()
-
+        return True if (self.removals or self.additions) else False
